### 2022-09-26 #4781 @martinlackner
Looking at the possible reviewers, maybe manmeetkaurbaxi, cosimameyer, hknd23 or pmyteh would be interested? They mention political science and/or computational social science - so there's some connection to abcvoting's topic (voting and computational social choice).
