+++ paper/paper.bib
@@ -5,7 +5,8 @@
   volume={35},
   number={6},
   pages={5448--5455},
-  year={2021}
+  year={2021},
+  doi = {10.1609/aaai.v35i6.16686},
 }
 
 @book{abcbook,
@@ -29,6 +30,7 @@
                Artificial Intelligence (IJCAI 2022)},
   pages     = {496--502},
   year      = {2022},
+  doi       = {10.24963/ijcai.2022/71},
 }
 
 @inproceedings{FairsteinVMG22,
@@ -50,7 +52,8 @@
   volume={36},
   number={5},
   pages={4892--4899},
-  year={2022}
+  year={2022},
+  doi = {10.1609/aaai.v36i5.20418},
 }
 
 @inproceedings{pjr17,
@@ -59,7 +62,8 @@
   Title = {Proportional justified representation},
   Year = {2017},
   pages = {670--676},
-  Xpublisher = {{AAAI} Press}
+  Xpublisher = {{AAAI} Press},
+  doi = {10.1609/aaai.v31i1.10611},
 }
 
 @article{justifiedRepresentation,
@@ -70,6 +74,7 @@
     volume="48",
     number="2",
     pages="461--485",
+    doi = {10.1609/aaai.v29i1.9324},
 }
 
 @inproceedings{pet-sko:laminar,
@@ -78,7 +83,8 @@
   Booktitle = "Proceedings of the 2020 ACM Conference on Economics and Computation (ACM-EC-2020)",
   year      = {2020},
   pages = {793--794},
-  note={Extended version at \url{https://arxiv.org/abs/1911.11747}}
+  url={https://arxiv.org/abs/1911.11747},
+  doi = {10.1145/3391403.3399465}
 }
 
 @inproceedings{MatteiW13,
@@ -86,7 +92,8 @@
   booktitle = "Proceedings of the 3rd International Conference on Algorithmic Decision Theory (ADT 2013)",
   pages = {259--270},
   title = {PrefLib: A Library for Preferences http: //www.preflib.org.},
-  year = 2013
+  year = 2013,
+  doi = {10.1007/978-3-642-41575-3_20}
 }
 
 @article{aij/guarantees,
@@ -97,6 +104,7 @@
   volume   = {288},
   pages    = {103366},
   issn     = {0004-3702},
+  doi      = {10.1016/j.artint.2020.103366},
 }
 
 @incollection{FSST-trends,
