### 2022-10-25 #4880 @editorialbot
[REVIEW]: abcvoting: A Python package for approval-based multi-winner voting rules
**Submitting author:** @martinlackner (<a href="http://orcid.org/0000-0003-2170-0770">Martin Lackner</a>)
**Repository:** https://github.com/martinlackner/abcvoting
**Branch with paper.md** (empty if default branch):
**Version:** v2.6.0
