### 2022-11-18 #4880 @martinlackner
Hi Christoper!
Thanks for your feedback!

- community guidelines: I will set up a brief text. There is not much to say, since this is a rather small project.
- license: abcvoting uses the [MIT license](https://github.com/martinlackner/abcvoting/blob/master/LICENSE).
- authorship: I also thought about that. Let me briefly sketch the situation: there are in total 9 people (including me) who contributed code to the Github repository. I am the maintainer and have contributed the majority of the code. I thought the best way of handling this is to mention everyone involved in the acknowledgements and limit the authors to those who actively develop abcvoting - which is just me. I am not particularly set on this way of handling authorship; if there is a standard procedure whom to include as authors I am happy to make changes.

Best,
Martin
