### 2022-11-18 5ec1bd1
Create CONTRIBUTING.md
