+++ CONTRIBUTING.md
@@ -1,0 +1,16 @@
+# Welcome to abcvoting
+
+Thank you for taking a closer look at our project! 
+
+We are very happy about all contributions! As this is a small project, 
+the best way to contribute is to simply get in contact with me (Martin Lackner, lackner@dbai.tuwien.ac.at) or to
+post an [issue](https://github.com/martinlackner/abcvoting/issues).
+
+Further comments:
+- If you would like to contribute but don't have a concrete idea, take a look the open [issues](https://github.com/martinlackner/abcvoting/issues) or
+contact me directly (Martin Lackner, lackner@dbai.tuwien.ac.at).
+- If you would like to contribute your favourite approval-based committee voting rules or axiom, you are very welcome to do so! 
+Take a look at [abcrules.py](https://github.com/martinlackner/abcvoting/blob/master/abcvoting/abcrules.py), where you find
+implementations of most standard rules. I am happy to help with integrating your new rule in the code (unit tests, etc.).
+- If you would like to contribute code, please clone the repository and make a pull request (PR) once you have a working version. 
+We can then discuss your code in detail. Feel free to get in touch before or while you are working on something.
