+++ paper/paper.bib
@@ -10,11 +10,12 @@
 }
 
 @book{abcbook,
-  title={Multi-winner voting with approval preferences},
+  title={Multi-Winner Voting with Approval Preferences},
   author={Lackner, Martin and Skowron, Piotr},
-  publisher={Springer. To appear},
-  year=2022,
-  url={https://arxiv.org/abs/2007.01795}
+  publisher={Springer International Publishing},
+  year=2023,
+  url={https://arxiv.org/abs/2007.01795},
+  doi={10.1007/978-3-031-09016-5},
 }
 
 @inproceedings{howtosample,
