### 2022-11-18 #4880 @martinfleis
This is what our guide tells about authorship:

> The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed.

Given the distribution of commits to the package, I would potentially consider adding _lumbric_ and _benjaminkrenn_ but will not oppose if you decide to keep it as it is now.
