### 2022-11-23 #4880 @martinlackner
Dear all,
I have contacted @lumbric and @benjaminkrenn and they agree to be listed as authors on the paper. I am happy about this solution regarding authorship, thanks @martinfleis for the suggestion.
I pushed an updated version of the paper to the repository (but I don't know how to update the linked version here in the thread).
Best,
Martin
