### 2022-11-28 5578f10
add alternative ways for equal shares to complete committees (increment budget, AV)
