+++ abcvoting/abcrules.py
@@ -37,6 +37,8 @@ MAIN_RULE_IDS = [
     "minimaxav",
     "lexminimaxav",
     "equal-shares",
+    "equal-shares-with-av-completion",
+    "equal-shares-with-increment-completion",
     "phragmen-enestroem",
     "consensus-rule",
     "trivial",
@@ -227,18 +229,32 @@ class Rule:
             self.compute_fct = compute_lexminimaxav
             self.algorithms = ("gurobi", "brute-force")
             self.resolute_values = self._RESOLUTE_VALUES_FOR_OPTIMIZATION_BASED_RULES
-        elif rule_id in ["rule-x", "equal-shares"]:
+        elif rule_id in ["rule-x", "equal-shares", "equal-shares-with-seqphragmen-completion"]:
             self.shortname = "Equal Shares"
-            self.longname = "Method of Equal Shares (aka Rule X)"
+            self.longname = "Method of Equal Shares (aka Rule X) with Phragmén phase"
             self.compute_fct = compute_equal_shares
             self.algorithms = ("float-fractions", "gmpy2-fractions", "standard-fractions")
             self.resolute_values = self._RESOLUTE_VALUES_FOR_SEQUENTIAL_RULES
-        elif rule_id in ["rule-x-without-phragmen-phase", "equal-shares-without-phragmen-phase"]:
-            self.shortname = "Equal Shares without Phragmén phase"
-            self.longname = (
-                "Method of Equal Shares (aka Rule X) without the Phragmén phase (second phase)"
-            )
-            self.compute_fct = functools.partial(compute_equal_shares, skip_phragmen_phase=True)
+        elif rule_id in ["equal-shares-with-av-completion"]:
+            self.shortname = "Equal Shares with AV completion"
+            self.longname = "Method of Equal Shares (aka Rule X) with AV completion"
+            self.compute_fct = functools.partial(compute_equal_shares, completion="av")
+            self.algorithms = ("float-fractions", "gmpy2-fractions", "standard-fractions")
+            self.resolute_values = self._RESOLUTE_VALUES_FOR_SEQUENTIAL_RULES
+        elif rule_id in ["equal-shares-with-increment-completion"]:
+            self.shortname = "Equal Shares with increment completion"
+            self.longname = "Method of Equal Shares (aka Rule X) with increment completion"
+            self.compute_fct = functools.partial(compute_equal_shares, completion="increment")
+            self.algorithms = ("float-fractions", "gmpy2-fractions", "standard-fractions")
+            self.resolute_values = (True,)  # this rule is ill-defined for resolute=False
+        elif rule_id in [
+            "rule-x-without-phragmen-phase",
+            "equal-shares-without-phragmen-phase",
+            "equal-shares-without-completion",
+        ]:
+            self.shortname = "Equal Shares without completion"
+            self.longname = "Method of Equal Shares (aka Rule X) without completion (second phase)"
+            self.compute_fct = functools.partial(compute_equal_shares, completion=None)
             self.algorithms = ("float-fractions", "gmpy2-fractions", "standard-fractions")
             self.resolute_values = self._RESOLUTE_VALUES_FOR_SEQUENTIAL_RULES
         elif rule_id == "phragmen-enestroem":
@@ -1881,7 +1897,9 @@ def compute_separable_rule(
     return committees
 
 
-def _separable_rule_algorithm(rule_id, profile, committeesize, resolute, max_num_of_committees):
+def _separable_rule_algorithm(
+    rule_id, profile, committeesize, resolute, max_num_of_committees=MAX_NUM_OF_COMMITTEES_DEFAULT
+):
     """
     Algorithm for separable rules (such as AV and SAV).
     """
[…]
