+++ tests/test_instances/instanceL0393.abc.yaml
@@ -164,6 +164,15 @@ compute:
 - rule_id: equal-shares
   result: [[2, 3, 5, 6, 7, 8]]
   resolute: false
+- rule_id: equal-shares-with-av-completion
+  result: [[2, 3, 5, 6, 7, 8]]
+  resolute: true
+- rule_id: equal-shares-with-av-completion
+  result: [[2, 3, 5, 6, 7, 8]]
+  resolute: false
+- rule_id: equal-shares-with-increment-completion
+  result: [[2, 3, 5, 6, 7, 8]]
+  resolute: true
 - rule_id: phragmen-enestroem
   result: [[2, 3, 5, 6, 7, 8]]
   resolute: true
