+++ tests/test_instances/instanceL0044.abc.yaml
@@ -125,6 +125,15 @@ compute:
 - rule_id: equal-shares
   result: [[0, 2, 4, 5]]
   resolute: false
+- rule_id: equal-shares-with-av-completion
+  result: [[0, 2, 4, 5]]
+  resolute: true
+- rule_id: equal-shares-with-av-completion
+  result: [[0, 2, 4, 5]]
+  resolute: false
+- rule_id: equal-shares-with-increment-completion
+  result: [[0, 2, 4, 5]]
+  resolute: true
 - rule_id: phragmen-enestroem
   result: [[0, 2, 4, 5]]
   resolute: true
