+++ tests/test_instances/instanceL0059.abc.yaml
@@ -120,6 +120,15 @@ compute:
 - rule_id: equal-shares
   result: [[2, 3, 4, 5]]
   resolute: false
+- rule_id: equal-shares-with-av-completion
+  result: [[2, 3, 4, 5]]
+  resolute: true
+- rule_id: equal-shares-with-av-completion
+  result: [[2, 3, 4, 5]]
+  resolute: false
+- rule_id: equal-shares-with-increment-completion
+  result: [[2, 3, 4, 5]]
+  resolute: true
 - rule_id: phragmen-enestroem
   result: [[0, 3, 4, 5]]
   resolute: true

[…]
