### 2022-12-21 cf25a6d
remove py3.7 from piptestfull (runtime too long)
