### 2023-01-16 3acaba9
update reference to book "Multi-Winner Voting with Approval Preferences" (now published)
